refactor: Move InputExtra class to a new file for better organization
